Move internal modules back to top level.
