+++ signac/__main__.py
@@ -30,12 +30,7 @@ else:
     READLINE = True
 
 from . import config, get_project, init_project
-from ._internal.utility import (
-    _add_verbosity_argument,
-    _print_err,
-    _query_yes_no,
-    _safe_relpath,
-)
+from ._utility import _add_verbosity_argument, _print_err, _query_yes_no, _safe_relpath
 from ._vendor.configobj import Section, flatten_errors
 from .diff import diff_jobs
 from .errors import (
