+++ signac/_utility.py
