+++ signac/diff.py
@@ -3,10 +3,7 @@
 # This software is licensed under the BSD 3-Clause License.
 """Compute diffs of state points."""
 
-from ._internal.utility import (
-    _dotted_dict_to_nested_dicts,
-    _nested_dicts_to_dotted_keys,
-)
+from ._utility import _dotted_dict_to_nested_dicts, _nested_dicts_to_dotted_keys
 
 
 def diff_jobs(*jobs):
