+++ signac/job.py
@@ -12,13 +12,13 @@ from json import JSONDecodeError
 from threading import RLock
 from typing import FrozenSet
 
-from ._internal.utility import _mkdir_p
 from ._synced_collections.backends.collection_json import (
     BufferedJSONAttrDict,
     JSONAttrDict,
     json_attr_dict_validator,
 )
 from ._synced_collections.errors import KeyTypeError
+from ._utility import _mkdir_p
 from .errors import DestinationExistsError, JobsCorruptedError
 from .h5store import H5StoreManager
 from .hashing import calc_id
